pyproject.toml
[tool.poetry]
name = "the_gratest_me"
version = "0.1.0"
description = ""
authors = ["Scareface <tektonikboy98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
